~/groq-mcp-server/requirements.txt
groq-sdk
mcp>=1.0.0
asyncio-mqtt
asyncio-task-queue